+++ apps/batches/wb/whisky/merge_execution.py
@@ -1,4 +1,4 @@
-from apps.batches.wb.whisky.wb_whisky_form import merged_whisky_data_preprocess
+from apps.batches.wb.whisky.merge_process import merged_whisky_data_preprocess
 from apps.batches.wb.common.enums import BatchType, CollectionName
 from apps.batches.libs.lib_firebase.firebase_funcs import save_to_firebase
 from apps.batches.wb.common import wb_libs_func
